requirements.txt
docker==4.2.0
Flask==1.1.1
Flask-RESTful==0.3.7
h5py==2.10.0
ipykernel==5.3.4
ipython==7.16.1
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
matplotlib==3.2.1
matroid==1.0.2
pandas==1.0.3
Pillow==5.4.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.3.0
seaborn==0.10.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
texttable==1.6.2
transformers==3.0.2
xmltodict==0.11.0
yacs==0.1.8